fix: (ab?)use Box to avoid burdening consumers of Log
Since rustc needs to know which impl to dispatch to based on Log type
params, consumers of Log need to know about Log's effective type at
compile-time. The reason that Log was generified to begin with was to
work with anything implementing the Write trait, and since Write is a
trait, it alone lacks a known size. As such, using Box to hold the
Write implementation in Log is a suitable compromise.